Improve Character Detail Screen

+++ app/src/main/java/com/noox/marvelheroes/core/extensions/UtilExtensions.kt
@@ -3,8 +3,8 @@ package com.noox.marvelheroes.core.extensions
 import android.content.res.Resources
 import android.util.TypedValue
 
-fun Float.dpToPx(): Int = TypedValue.applyDimension(
+fun Float.dpToPx(): Float = TypedValue.applyDimension(
     TypedValue.COMPLEX_UNIT_DIP, this, Resources.getSystem().displayMetrics
-).toInt()
+)
 
-fun Int.dpToPx(): Int = this.toFloat().dpToPx()
+fun Int.dpToPx(): Int = this.toFloat().dpToPx().toInt()
